fix: GH-69 troubleshooting iOS running migration on startup, add button to admin page to trigger migration, move other migration and update so it doesn't require the vuex store
I'm still not convinced I've solved the 'the request has not finished' error that we were seeing on Safari. It seems like Safari will happily run the migration once the page has been alive for a bit but trying to run it on startup is when we'd see the error, so it feels like a timing issue. I hope this change where we lock out other indexedDB related processes fixes it